% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpoly.R
\name{bw.lp1D}
\alias{bw.lp1D}
\title{select bandwidth for one-dimensional local linear smoother}
\usage{
bw.lp1D(x, y, weight, kernel, degree, method = "cv", K = 5, H = NULL)
}
\arguments{
\item{x}{a vector of observed values of the predictor}

\item{y}{a vector of observed values of the response}

\item{weight}{a vector of real numbers obtaining the weight for each observations}

\item{kernel}{a kernel from the package locpol}

\item{degree}{the degree of local polynomials}

\item{method}{selection method, supported is 'cv'}

\item{K}{the number of folds, a parameter for CV method}

\item{H}{candidate values of the bandwidth; if NULL, then automathcally generated}
}
\value{
the selected bandwidth
}
\description{
select bandwidth for one-dimensional local linear smoother
}
