% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpoly.R
\name{compute.bw.1D}
\alias{compute.bw.1D}
\title{select bandwidth for one-dimensional local linear smoother}
\usage{
compute.bw.1D(x, y, method = c("cv", "plug.in", "thumb"), weig, kernel,
  deg)
}
\arguments{
\item{x}{a vector of observed values of the predictor}

\item{y}{a vector of observed values of the response}

\item{method}{selection method}

\item{weig}{a vector of real numbers obtaining the weight for each observations}

\item{kernel}{a kernel from the package locpol}

\item{deg}{the degree of local polynomials}
}
\value{
the selected bandwidth
}
\description{
select bandwidth for one-dimensional local linear smoother
}
\keyword{internal}
