% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.partition}
\alias{cv.partition}
\title{the function mimic the cvpartition in MATLAB}
\usage{
cv.partition(n, K)
}
\arguments{
\item{n}{sample size}

\item{K}{the number of cv folders}
}
\value{
a list of the follow elements
\item{training}{a list of \code{K} elements. Each element is a vector of indices of samples for training}
\item{test}{a list of \code{K} elements. Each element is a vector of indices of samples for test}
\item{num.test.sets}{an integer signifying the number of test sets}
}
\description{
the function mimic the cvpartition in MATLAB
}
