% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.sp.R
\name{estimate.theta}
\alias{estimate.theta}
\title{Estimate the parameters of the correlation structure}
\usage{
estimate.theta(
  Lt,
  Ly,
  D,
  var.hat,
  mu.hat,
  method,
  rho,
  weig = NULL,
  theta.lb = NULL,
  theta.ub = NULL,
  h = 0.1,
  domain = c(0, 1),
  pfunc = NULL,
  theta0 = NULL
)
}
\arguments{
\item{D}{the number of parameters}

\item{weig}{a vector of length(Lt)}

\item{sig0.hat}{the estimate of variance of measure error}
}
\description{
Estimate the parameters of the correlation structure
}
\keyword{internal}
