% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpoly.R
\name{loclin1D}
\alias{loclin1D}
\title{one-dimensional local linear smoother}
\usage{
loclin1D(x, y, newx, bw = "cv", weig = rep(1, length(y)),
  kernel = EpaK, deg = 1)
}
\arguments{
\item{x}{a vector of observed values of the predictor}

\item{y}{a vector of observed values of the response}

\item{newx}{a vector of new values of the predictor}

\item{bw}{the bandwidth, either a numeric or an element from \code{c('cv','plug.in','thumb')}}

\item{weig}{a vector of real numbers obtaining the weight for each observations}

\item{kernel}{a kernel from the package locpol}

\item{deg}{the degree of local polynomials}
}
\value{
a list of the following elements
\item{bw}{the selected bandwidth or input bandwidth}
\item{x}{the original input}
\item{y}{the original input}
\item{weig}{the original input}
\item{kernel}{the original input}
\item{deg}{the original input}
}
\description{
one-dimensional local linear smoother
}
