% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpoly.R
\name{lp1D}
\alias{lp1D}
\title{one-dimensional local polynomial smoother}
\usage{
lp1D(
  x,
  y,
  h = NULL,
  newx = x,
  degree = 1,
  weight = NULL,
  kernel = "epanechnikov",
  sorted = F
)
}
\arguments{
\item{x}{a vector of observed values of the predictor}

\item{y}{a vector of observed values of the response}

\item{h}{the bandwidth, either a numeric or NULL; if NULL, the automatically selected by five-fold CV}

\item{newx}{a vector of new values of the predictor; set to \code{x} by default}

\item{degree}{the degree of local polynomials, nonnegative integer}

\item{weight}{a vector of real numbers obtaining the weight for each observations; if NULL, then equal weight}

\item{kernel}{a kernel from the package locpol, supported are 'epanechnikov', "rectangular", "triangular", "quartic", "triweight", "tricube", "cosine", "gauss", "logistic", "sigmoid" and "silverman"}

\item{sorted}{indicator of whether the array x and newx are sorted in increasing order}
}
\value{
a list of the following elements
\item{h}{the selected bandwidth or input bandwidth}
\item{x}{the original input}
\item{y}{the original input}
\item{weight}{the original input}
\item{kernel}{the original input}
\item{degree}{the original input}
\item{yhat}{the estimated response at \code{newx}}
}
\description{
one-dimensional local polynomial smoother
}
