% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{meanfunc}
\alias{meanfunc}
\title{Estimate the mean function from functional data/snippets}
\usage{
meanfunc(
  t,
  y,
  newt = NULL,
  method = c("PACE", "FOURIER"),
  tuning = "cv",
  weig = NULL,
  ...
)
}
\arguments{
\item{t}{a list of vectors (for irregular design) or a vector (for regular design) containing time points of observations for each individual. Each vector should be in ascending order}

\item{y}{a list of vectors (for irregular design) or a matrix (for regular design) containing the observed values at \code{t}. If it is a matrix, the columns correspond to the time points in the vector \code{t}}

\item{newt}{a list of vectors or a vector containing time points of observations to be evaluated. If NULL, then newt is treated as t}

\item{method}{estimation method, 'PACE' or 'FOURIER'}

\item{tuning}{tuning method to select possible tuning parameters}

\item{...}{other parameters required depending on the \code{method} and \code{tuning}; see details}
}
\value{
an object of the class 'meanfunc' containing necessary information to predict the mean function
}
\description{
Estimate the mean function from functional data/snippets
}
\details{
\itemize{
        \item{When \code{method='PACE'}, additional parameters \code{kernel} and \code{deg} can be provided. \code{bw} as a scalar is optional. When \code{bw} is provided, the bandwidth is set to \code{bw}}
        \item{When \code{method='FOURIER'}, additional parameters \code{q},\code{rho},\code{ext} and \code{domain} are optional. If they are not provided, then they will be deduced from data or selected by the specified \code{tuning} method.}
    }
}
\examples{
mu <- function(s) sin(2*pi*s)
D <- synfd::sparse.fd(mu=mu, X=synfd::gaussian.process(), n=100, m=5)
mu.obj <- meanfunc(D$t,D$y,newt=NULL,method='PACE',
                tuning='cv',weig=NULL,kernel='gauss',deg=1)
# equivalent to
# mu.obj <- meanfunc(D$t,D$y)

# plot the object
plot(mu.obj)
}
\references{
\insertRef{Lin2020}{mcfda}

\insertRef{Yao2005}{mcfda}
}
