% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{plot.meanfunc}
\alias{plot.meanfunc}
\title{plot the estimated mean function}
\usage{
\method{plot}{meanfunc}(meanfunc.obj, ...)
}
\arguments{
\item{meanfunc.obj}{the object obtained by calling \code{meanfunc}}

\item{...}{other parameters passed to \code{plot}}
}
\description{
plot the estimated mean function
}
