% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{predict.meanfunc}
\alias{predict.meanfunc}
\title{predict mean functions at new locations}
\usage{
\method{predict}{meanfunc}(meanfunc.obj, newt)
}
\arguments{
\item{meanfunc.obj}{the object obtained by calling \code{mean.func}}

\item{newt}{a vector or a list of vectors of real numbers}
}
\value{
the estimated mean function evaluated at \code{newt}. It has the same format of \code{newt}
}
\description{
predict mean functions at new locations
}
