% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{regular.grid}
\alias{regular.grid}
\title{generate a regular grid on a domain}
\usage{
regular.grid(m = 100, domain = c(0, 1), h = 1/(2 * m))
}
\arguments{
\item{m}{the number of points}

\item{domain}{the domain}

\item{h}{the margin of the first and last points to the boundaries of the domain}
}
\value{
a vector representing the grid
}
\description{
generate a regular grid on a domain
}
\examples{
grid <- regular.grid()
grid <- regular.grid(50)
}
