% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rep.row}
\alias{rep.row}
\title{replicate a row vector into a matrix}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{the row vector to be replicated}

\item{n}{replicate the row vector n times}
}
\value{
a matrix of the dimension of \code{n} by \code{length(x)}
}
\description{
replicate a row vector into a matrix
}
