% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2.R
\name{sigma2}
\alias{sigma2}
\title{estimate the variance of noise}
\usage{
sigma2(t, y, h = NULL)
}
\arguments{
\item{t}{a list of vectors containing time points for each individual. Each vector is sorted in ascending order.}

\item{y}{a list of vectors containing observed values at \code{t}}

\item{h}{the bandwidth}
}
\value{
the estimated variance of the measurement error
}
\description{
estimate the variance of noise
}
