% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var.func}
\alias{var.func}
\title{Estimate the variance function}
\usage{
var.func(Lt, Ly, newt = NULL, sig2 = NULL, method = c("PACE",
  "FOURIER"), mu = NULL, weig = NULL, ...)
}
\arguments{
\item{Lt}{a list (for irregular design) or a vector (for regular design) of observations in time domain}

\item{Ly}{a list (for irregular design) or a matrix (for regular design) of observations corresponding to \code{Lt}}

\item{newt}{a set of locations where the variance function is estimated. If NULL, then set to \code{Lt}}

\item{sig2}{variance of noise}

\item{method}{estimation method, 'PACE' or 'FOURIER'}

\item{mu}{an object generated by \code{meanfunc}.}

\item{weig}{a vector of length(Lt). If NULL, then determined by the algorithm.}

\item{...}{other parameters required depending on the \code{method} and \code{tuning}; see details}
}
\description{
Estimate the variance function
}
\details{
\itemize{
        \item{When \code{method='PACE'}, additional parameters \code{kernel} and \code{deg} can be provided. \code{bw} as a scalar is optional. When \code{bw} is provided, the bandwidth is set to \code{bw}}
        \item{When \code{method='FOURIER'}, additional parameters \code{q},\code{rho},\code{ext} and \code{domain} are optional. If they are not provided, then they will be deduced from data or selected by the specified \code{tuning} method.}
    }
}
